Update core to MSAL.Net (Microsoft.Identity)

+++ Vault/Library/FileTokenCache.cs
@@ -1,19 +1,17 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved. 
 // Licensed under the MIT License. See License.txt in the project root for license information. 
 
-using Microsoft.IdentityModel.Clients.ActiveDirectory;
+using Microsoft.Identity.Client;
 using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
-using System.Security.AccessControl;
 using System.Security.Cryptography;
-using System.Text;
 using System.Threading.Tasks;
 
 namespace Microsoft.Vault.Library
 {
-    public class FileTokenCache : TokenCache
+    public class FileTokenCache
     {
         public string FileName;
         private static readonly object FileLock = new object();
@@ -22,16 +20,13 @@ namespace Microsoft.Vault.Library
 
         /// <summary>
         /// Initializes the cache against a local file.
-        /// If the file is already present, it loads its content in the ADAL cache
+        /// If the file is already present, it loads its content in the MSAL cache
         /// </summary>
         /// <param name="domainHint">For example: microsoft.com or gme.gbl</param>
         public FileTokenCache(string domainHint)
         {
             FileName = Environment.ExpandEnvironmentVariables(string.Format(Consts.VaultTokenCacheFileName, domainHint));
             Directory.CreateDirectory(Path.GetDirectoryName(FileName));
-            this.AfterAccess = AfterAccessNotification;
-            this.BeforeAccess = BeforeAccessNotification;
-            BeforeAccessNotification(null);
         }
 
         /// <summary>
@@ -75,51 +70,79 @@ namespace Microsoft.Vault.Library
                 File.Delete(newName);
             }
             File.Move(FileName, newName);
-
             FileName = newName;
-            this.AfterAccess = AfterAccessNotification;
-            this.BeforeAccess = BeforeAccessNotification;
-            BeforeAccessNotification(null);
         }
-
 
         /// <summary>
         /// Empties the persistent store
         /// </summary>
-        public override void Clear()
+        public void Clear()
         {
-            base.Clear();
-            File.Delete(FileName);
-        }
-
-        /// <summary>
-        /// Triggered right before ADAL needs to access the cache
-        /// Reload the cache from the persistent store in case it changed since the last access
-        /// </summary>
-        /// <param name="args"></param>
-        void BeforeAccessNotification(TokenCacheNotificationArgs args)
-        {
-            lock (FileLock)
+            if (File.Exists(FileName))
             {
-                this.Deserialize(File.Exists(FileName) ? ProtectedData.Unprotect(File.ReadAllBytes(FileName), null, DataProtectionScope.CurrentUser) : null);
+                File.Delete(FileName);
             }
         }
 
         /// <summary>
-        /// Triggered right after ADAL accessed the cache
+        /// Configures the token cache for an MSAL client application
+        /// </summary>
+        /// <param name="tokenCache">The MSAL token cache to configure</param>
+        public void ConfigureTokenCache(ITokenCache tokenCache)
+        {
+            tokenCache.SetBeforeAccess(BeforeAccessNotification);
+            tokenCache.SetAfterAccess(AfterAccessNotification);
+        }
+
+        /// <summary>
+        /// Triggered right before MSAL needs to access the cache
+        /// Reload the cache from the persistent store in case it changed since the last access
         /// </summary>
         /// <param name="args"></param>
-        void AfterAccessNotification(TokenCacheNotificationArgs args)
+        private void BeforeAccessNotification(TokenCacheNotificationArgs args)
+        {
+            lock (FileLock)
+            {
+                if (File.Exists(FileName))
+                {
+                    try
+                    {
+                        byte[] encryptedData = File.ReadAllBytes(FileName);
+                        byte[] data = ProtectedData.Unprotect(encryptedData, null, DataProtectionScope.CurrentUser);
+                        args.TokenCache.DeserializeMsalV3(data);
+                    }
+                    catch (Exception)
+                    {
+                        // If decryption fails, clear the cache and start fresh
+                        Clear();
+                    }
+                }
+            }
+        }
+
+        /// <summary>
+        /// Triggered right after MSAL accessed the cache
+        /// </summary>
+        /// <param name="args"></param>
+        private void AfterAccessNotification(TokenCacheNotificationArgs args)
         {
             // if the access operation resulted in a cache update
-            if (this.HasStateChanged)
+            if (args.HasStateChanged)
             {
                 lock (FileLock)
                 {
-                    // reflect changes in the persistent store
-                    File.WriteAllBytes(FileName, ProtectedData.Protect(this.Serialize(), null, DataProtectionScope.CurrentUser));
-                    // once the write operation took place, restore the HasStateChanged bit to false
-                    this.HasStateChanged = false;
+                    try
+                    {
+                        // reflect changes in the persistent store
+                        byte[] data = args.TokenCache.SerializeMsalV3();
+                        byte[] encryptedData = ProtectedData.Protect(data, null, DataProtectionScope.CurrentUser);
+                        File.WriteAllBytes(FileName, encryptedData);
+                    }
+                    catch (Exception)
+                    {
+                        // If encryption fails, don't crash the application
+                        // Log the error if logging is available
+                    }
                 }
             }
         }
